requirements.txt
altair==5.2.0
altgraph==0.17.4
annotated-types==0.6.0
appdirs==1.4.4
asttokens==2.4.1
attrs==23.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.7.0
blis==0.7.11
breadability==0.1.20
bs4==0.0.2
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chromedrivermanager==0.0.1
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contourpy==1.2.0
cssselect==1.2.0
cycler==0.12.1
cymem==2.0.8
dateparser==1.2.0
decorator==5.1.1
demjson3==3.0.6
distlib==0.3.8
docopt==0.6.2
emoji==2.12.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz#sha256=65cea113d7509d5ab515583bbcbc9a31b4157bebbb73133d4110d2945476eaa4
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.7.0/es_core_news_sm-3.7.0-py3-none-any.whl#sha256=61e6e5530941f5880166855f09f60d7e6ba79ec1e8e45f96244bdb1eb169eb1d
et-xmlfile==1.1.0
executing==2.1.0
extra-streamlit-components==0.1.71
facebook-scraper==0.2.59
fake-useragent==1.5.1
filelock==3.16.1
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.47.2
funcy==2.0
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.41
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
h11==0.14.0
httplib2==0.22.0
idna==3.6
importlib-metadata==7.0.1
instaloader==4.11
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpickle==3.3.0
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langcodes==3.3.0
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
murmurhash==1.0.10
networkx==3.3
nltk==3.8.1
numexpr==2.10.1
numpy==1.26.3
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
parse==1.20.2
parso==0.8.4
pefile==2023.2.7
pillow==10.2.0
platformdirs==4.3.6
plotly==5.18.0
preshed==3.0.9
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.2
pure_eval==0.2.3
pyarrow==14.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==23.12.11
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
pydub==0.25.1
pyee==11.1.1
Pygments==2.17.2
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.6
PyJWT==2.8.0
pyLDAvis==3.4.1
pyparsing==3.1.1
PyPDF2==3.0.1
pypiwin32==223
pyppeteer==2.0.0
pyquery==2.0.1
pyreadstat==1.2.7
PySocks==1.7.1
pytesseract==0.3.10
python-barcode==0.15.1
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2023.3.post1
pyvis==0.3.2
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-html==0.10.0
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
schedule==1.2.1
scikit-learn==1.4.0
scipy==1.11.4
seaborn==0.13.2
selenium==4.21.0
sentiment-analysis-spanish==0.0.25
setuptools==69.0.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.1
sr==0.1.0
srsly==2.4.8
stack-data==0.6.3
stopwords==1.0.1
streamlit==1.30.0
streamlit-authenticator==0.3.2
streamlit-wordcloud-v2==0.4.0
sumy==0.11.0
tenacity==8.2.3
textblob==0.18.0.post0
thinc==8.2.2
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.3
trio==0.25.1
trio-websocket==0.11.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.20
validators==0.22.0
virtualenv==20.26.5
w3lib==2.2.1
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.13
weasel==0.3.4
webdriver-manager==4.0.1
websockets==10.4
Werkzeug==3.0.3
wheel==0.42.0
wordcloud==1.9.3
wsproto==1.2.0
zipp==3.17.0
